+++ src/test/java/com/sparta/ticketauction/domain/user/util/UserUtil.java
@@ -1,7 +1,10 @@
-package com.sparta.ticketauction.domain.user;
+package com.sparta.ticketauction.domain.user.util;
 
 import java.time.LocalDate;
 
+import org.springframework.security.crypto.password.PasswordEncoder;
+
+import com.sparta.ticketauction.domain.user.entity.User;
 import com.sparta.ticketauction.domain.user.request.UserCreateRequest;
 
 public class UserUtil {
@@ -12,6 +15,10 @@ public class UserUtil {
 	public static final String NICKNAME = "두루미";
 	public static final String PHONE_NUMBER = "010-1234-5678";
 	public static final LocalDate BIRTH = LocalDate.of(1990, 1, 1);
+
+	public static User createUser(PasswordEncoder passwordEncoder) {
+		return getUserCreateRequest().toEntity(passwordEncoder);
+	}
 
 	public static UserCreateRequest getUserCreateRequest() {
 		return new UserCreateRequest(
